feat:jwt auth in all comment, experience and users routes with userId based updates and delete of each

+++ src/features/comments/comment.controller.ts
@@ -1,6 +1,11 @@
 import type { RequestHandler } from "express";
 import type CommentService from "./comment.service.js";
-import { COMMENT_MESSAGE, HTTP_STATUS } from "../../constants/constants.js";
+import {
+  AUTH_MESSAGE,
+  COMMENT_MESSAGE,
+  HTTP_STATUS,
+} from "../../constants/constants.js";
+import AppError from "../../utils/error.handler.js";
 
 export default class CommentController {
   constructor(private commentService: CommentService) {}
@@ -31,8 +36,14 @@ export default class CommentController {
     });
   };
 
-  update: RequestHandler = async (req, res) => {
+  update: RequestHandler = async (req, res, next) => {
     const commentId = Number(req.params.commentId);
+    const userIdObj = await this.commentService.findUserid(commentId);
+    if (req.userId !== userIdObj?.userId) {
+      return next(
+        new AppError(HTTP_STATUS.FORBIDDEN, AUTH_MESSAGE.NOT_PERMITTED),
+      );
+    }
     const comment = req.body;
     return res.status(HTTP_STATUS.OK).json({
       success: true,
@@ -41,8 +52,14 @@ export default class CommentController {
     });
   };
 
-  delete: RequestHandler = async (req, res) => {
+  delete: RequestHandler = async (req, res, next) => {
     const commentId = Number(req.params.commentId);
+    const userIdObj = await this.commentService.findUserid(commentId);
+    if (req.userId !== userIdObj?.userId) {
+      return next(
+        new AppError(HTTP_STATUS.FORBIDDEN, AUTH_MESSAGE.NOT_PERMITTED),
+      );
+    }
     await this.commentService.delete(commentId);
     return res.status(HTTP_STATUS.NO_CONTENT).send();
   };
