+++ src/features/users/user.controller.ts
@@ -1,7 +1,11 @@
 import type { NextFunction, Request, Response } from "express";
 import UserService from "./user.service.js";
 import { catchAsync } from "../../utils/catch.async.js";
-import { HTTP_STATUS, USER_MESSAGE } from "../../constants/constants.js";
+import {
+  AUTH_MESSAGE,
+  HTTP_STATUS,
+  USER_MESSAGE,
+} from "../../constants/constants.js";
 import AppError from "../../utils/error.handler.js";
 
 export default class UserController {
@@ -20,11 +24,13 @@ export default class UserController {
 
   //get user by id
   getUserById = catchAsync(
-    async (req: Request, res: Response,next:NextFunction) => {
+    async (req: Request, res: Response, next: NextFunction) => {
       const id = Number(req.params.id);
       const data = await this.userService.getUserById(id);
       if (!data) {
-        return next(new AppError(HTTP_STATUS.NOT_FOUND, USER_MESSAGE.USER_FETCH_FAIL(id)));
+        return next(
+          new AppError(HTTP_STATUS.NOT_FOUND, USER_MESSAGE.USER_FETCH_FAIL(id)),
+        );
       }
       return res.status(HTTP_STATUS.OK).json({
         success: true,
@@ -34,21 +40,24 @@ export default class UserController {
   );
 
   //get all users
-  getAllUsers = catchAsync(
-    async (req: Request, res: Response) => {
-      const data = await this.userService.getAllUsers();
-      
-      return res.status(HTTP_STATUS.OK).json({
-        success: true,
-        data: data,
-      });
-    },
-  );
+  getAllUsers = catchAsync(async (req: Request, res: Response) => {
+    const data = await this.userService.getAllUsers();
+
+    return res.status(HTTP_STATUS.OK).json({
+      success: true,
+      data: data,
+    });
+  });
 
   //update User by id
   updatedUser = catchAsync(
-    async (req: Request, res: Response) => {
+    async (req: Request, res: Response, next: NextFunction) => {
       const id = Number(req.params.id);
+      if (req.userId !== id) {
+        return next(
+          new AppError(HTTP_STATUS.FORBIDDEN, AUTH_MESSAGE.NOT_PERMITTED),
+        );
+      }
       const data = req.body;
       const updatedUser = await this.userService.updateUser(id, data);
       console.log(updatedUser);
@@ -62,8 +71,13 @@ export default class UserController {
 
   //delete user
   deletedUser = catchAsync(
-    async (req: Request, res: Response) => {
+    async (req: Request, res: Response, next: NextFunction) => {
       const id = Number(req.params.id);
+      if (req.userId !== id) {
+        return next(
+          new AppError(HTTP_STATUS.FORBIDDEN, AUTH_MESSAGE.NOT_PERMITTED),
+        );
+      }
       await this.userService.deleteUser(Number(id));
       res.status(HTTP_STATUS.NO_CONTENT).send();
     },
